requirements.txt
aiofiles==24.1.0
httpx==0.27.2
python-dotenv==1.0.1
setuptools==75.2.0
twine==5.1.1
